+++ src/Commands/EncryptCommand.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace pxgamer\Arionum\Console\Commands;
+namespace pxgamer\ArionumCLI\Commands;
 
-use pxgamer\Arionum\Console\BaseCommand;
+use pxgamer\ArionumCLI\BaseCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use function strlen;
@@ -10,7 +10,7 @@ use function strlen;
 /**
  * Class EncryptCommand
  */
-class EncryptCommand extends BaseCommand
+final class EncryptCommand extends BaseCommand
 {
     protected function configure(): void
     {
@@ -24,10 +24,10 @@ class EncryptCommand extends BaseCommand
     /**
      * @param InputInterface  $input
      * @param OutputInterface $output
-     * @return int|null|void
+     * @return void
      * @throws \Exception
      */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): void
     {
         parent::execute($input, $output);